# openapi.yaml
openapi: 3.1.0
info:
  title: viguno
  description: Phenotype/disease for VarFish
  contact:
    name: Manuel Holtgrewe
    email: manuel.holtgrewe@bih-charite.de
  license:
    name: Apache-2.0
  version: 0.3.2
paths:
  /hpo/genes:
    get:
      tags:
      - hpo_genes
      summary: Query for genes in the HPO database.
      operationId: hpo_genes
      parameters:
      - name: gene_id
        in: query
        description: The gene ID to search for.
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: gene_symbol
        in: query
        description: The gene symbol to search for.
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: match_
        in: query
        description: The match mode.
        required: false
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Match'
      - name: max_results
        in: query
        description: Maximal number of results to return.
        required: false
        schema:
          type: integer
          minimum: 0
      - name: hpo_terms
        in: query
        description: Whether to include HPO terms.
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: The query was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
  /hpo/omims:
    get:
      tags:
      - hpo_omims
      summary: Query for OMIM diseases in the HPO database.
      operationId: hpo_omims
      parameters:
      - name: omim_id
        in: query
        description: The OMIM ID to search for.
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: name
        in: query
        description: The disease name to search for.
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: match_
        in: query
        description: The match mode, default is `Match::Exact`.
        required: false
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Match'
      - name: ignore_case
        in: query
        description: Whether case is insentivie, default is `false`.
        required: false
        schema:
          type:
          - boolean
          - 'null'
      - name: max_results
        in: query
        description: Maximal number of results to return.
        required: false
        schema:
          type: integer
          minimum: 0
      - name: hpo_terms
        in: query
        description: Whether to include HPO terms.
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: The query was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
  /hpo/sim/term-gene:
    get:
      tags:
      - hpo_sim::term_gene
      summary: |-
        Query for similarity between a set of terms to each entry in a
        list of genes.
      operationId: hpo_sim_term_gene
      parameters:
      - name: terms
        in: query
        description: Set of terms to use as query.
        required: true
        schema:
          type: array
          items:
            type: string
      - name: gene_ids
        in: query
        description: The set of ids for genes to use as "database".
        required: false
        schema:
          type:
          - array
          - 'null'
          items:
            type: string
      - name: gene_symbols
        in: query
        description: The set of symbols for genes to use as "database".
        required: false
        schema:
          type:
          - array
          - 'null'
          items:
            type: string
      responses:
        '200':
          description: The query was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
  /hpo/sim/term-term:
    get:
      tags:
      - hpo_sim::term_term
      summary: Query for pairwise term similarity.
      description: |-
        In the case of Resnik, this corresponds to `IC(MICA(t_1, t_2))`.

        # Errors

        In the case that there is an error running the server.
      operationId: hpo_sim_term_term
      parameters:
      - name: lhs
        in: query
        description: The one set of HPO terms to compute similarity for.
        required: true
        schema:
          type: array
          items:
            type: string
      - name: rhs
        in: query
        description: The second set of HPO terms to compute similarity for.
        required: true
        schema:
          type: array
          items:
            type: string
      - name: ic_base
        in: query
        description: What should information content be based on.
        required: false
        schema:
          $ref: '#/components/schemas/IcBasedOn'
      - name: similarity
        in: query
        description: The similarity method to use.
        required: false
        schema:
          $ref: '#/components/schemas/SimilarityMethod'
      - name: combiner
        in: query
        description: The score combiner.
        required: false
        schema:
          $ref: '#/components/schemas/ScoreCombiner'
      responses:
        '200':
          description: The query was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
  /hpo/terms:
    get:
      tags:
      - hpo_terms
      summary: Query for terms in the HPO database.
      description: |-
        # Errors

        In the case that there is an error running the server.
      operationId: hpo_terms
      parameters:
      - name: term_id
        in: query
        description: The term ID to search for.
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: name
        in: query
        description: The term name to search for.
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: max_results
        in: query
        description: Maximal number of results to return.
        required: false
        schema:
          type: integer
          minimum: 0
      - name: genes
        in: query
        description: Whether to include genes.
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: The query was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result_'
components:
  schemas:
    Gene:
      type: object
      description: Struct for storing gene information in the result.
      required:
      - entrez_id
      - gene_symbol
      properties:
        entrez_id:
          type: integer
          format: int32
          description: The NCBI gene ID.
          minimum: 0
        gene_symbol:
          type: string
          description: The gene symbol.
        hgnc_id:
          type:
          - string
          - 'null'
          description: The HGNC ID.
    HpoTerm:
      type: object
      description: Struct for loading an HPO term from JSON.
      required:
      - term_id
      properties:
        term_id:
          type: string
          description: The term ID.
        term_name:
          type:
          - string
          - 'null'
          description: The term name (optional).
    IcBasedOn:
      type: string
      description: |-
        Enum for representing the information content kind.

        We replicate what is in the `hpo` create so we can put them on the command line and use
        them in HTTP queries more easily.
      enum:
      - gene
      - omim
    Match:
      type: string
      description: Specify how to perform query matches in the API calls.
      enum:
      - exact
      - prefix
      - suffix
      - contains
    Query:
      type: object
      title: HpoSimTermGeneQuery
      description: The performed query.
      required:
      - terms
      - genes
      properties:
        terms:
          type: array
          items:
            $ref: '#/components/schemas/HpoTerm'
          description: The query HPO terms.
        genes:
          type: array
          items:
            $ref: '#/components/schemas/Gene'
          description: The gene list to score.
    ResponseQuery:
      type: object
      title: HpoSimTermTermQuery
      description: |-
        Request as sent together with the response.

        The difference is that the `lhs` and `rhs` fields are replaced by vecs.
      required:
      - lhs
      - rhs
      properties:
        lhs:
          type: array
          items:
            type: string
          description: The one set of HPO terms to compute similarity for.
        rhs:
          type: array
          items:
            type: string
          description: The second set of HPO terms to compute similarity for.
        ic_base:
          $ref: '#/components/schemas/IcBasedOn'
          description: What should information content be based on.
        similarity:
          $ref: '#/components/schemas/SimilarityMethod'
          description: The similarity method to use.
        combiner:
          $ref: '#/components/schemas/ScoreCombiner'
          description: The score combiner.
    Result:
      type: object
      title: HpoSimTermGeneResult
      description: Result container data structure.
      required:
      - version
      - query
      - result
      properties:
        version:
          $ref: '#/components/schemas/Version'
          description: Version information.
        query:
          $ref: '#/components/schemas/Query'
          description: The original query records.
        result:
          type: array
          items:
            $ref: '#/components/schemas/ResultEntry'
          description: The resulting records for the scored genes.
    ResultEntry:
      type: object
      title: HpoSimTermGeneResultEntry
      description: Store score for a record with information on individual terms.
      required:
      - gene_symbol
      - raw_score
      properties:
        gene_symbol:
          type: string
          description: The gene symbol.
        raw_score:
          type: number
          format: float
          description: The raw Phenomizer score.
        terms:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/TermDetails'
          description: Details on individual terms.
    ResultGene:
      type: object
      description: Representation of a gene.
      required:
      - ncbi_gene_id
      - gene_symbol
      properties:
        ncbi_gene_id:
          type: integer
          format: int32
          description: The HPO ID.
          minimum: 0
        gene_symbol:
          type: string
          description: The description.
        hgnc_id:
          type:
          - string
          - 'null'
          description: The HGNC ID.
    ResultHpoTerm:
      type: object
      description: Representation of an HPO term.
      required:
      - term_id
      - name
      properties:
        term_id:
          type: string
          description: The HPO ID.
        name:
          type: string
          description: The term name.
    Result_:
      type: object
      title: HpoTermsResult
      description: Container for the result.
      required:
      - version
      - query
      - result
      properties:
        version:
          $ref: '#/components/schemas/Version'
          description: Version information.
        query:
          $ref: '#/components/schemas/Query'
          description: The original query records.
        result:
          type: array
          items:
            $ref: '#/components/schemas/ResultEntry'
          description: The resulting records for the scored genes.
    ScoreCombiner:
      type: string
      description: |-
        Representation of the standard combiners from HPO.

        We replicate what is in the `hpo` create so we can put them on the command line and use
        them in HTTP queries more easily.
      enum:
      - fun-sim-avg
      - fun-sim-max
      - bma
    SimilarityMethod:
      type: string
      description: |-
        Enum for representing similarity method to use.

        We replicate what is in the `hpo` create so we can put them on the command line and use
        them in HTTP queries more easily.
      enum:
      - distance-gene
      - graph-ic
      - information-coefficient
      - jc
      - lin
      - mutation
      - relevance
      - resnik
    TermDetails:
      type: object
      title: HpoSimTermGeneTermDetails
      description: Detailed term scores.
      required:
      - term_gene
      - score
      properties:
        term_query:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/HpoTerm'
            description: The query HPO term.
        term_gene:
          $ref: '#/components/schemas/HpoTerm'
          description: The gene's HPO term.
        score:
          type: number
          format: float
          description: The similarity score.
    Version:
      type: object
      description: Version information that is returned by the HTTP server.
      required:
      - hpo
      - viguno
      properties:
        hpo:
          type: string
          description: Version of the HPO.
        viguno:
          type: string
          description: Version of the `viguno` package.